Add more comments, rename classes, move functions adaptors to separate file.
